backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic>=2.7.0
python-multipart==0.0.6
httpx==0.25.0
passlib[bcrypt]==1.7.4
python-jose
pymysql
pydantic-settings==2.1.0
openai>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
python-dotenv>=1.0.0
elevenlabs==2.16.0
PyYAML
alembic
pytest-cov
boto3
bcrypt==4.0.1
